Image inpainting using patch-based and texture synthesis algorithms .
Image  inpainting is a art of missing value or a data in an image. The purpose of image inpainting is to reconstruct missing regions
which is visible for human eyes. Image inpainting is the process of reconstructing lost part of images based on the background
information. Image inpainting is a technique for restoring damaged old photographs and removing undesired objects from an image.